refactor(PN-10908): change package name
